Rename mirror types to generate types.
